Orginazing the tests
The test files now located in:
account.tests.test_profile
account.tests.test_professional
account.tests.test_client
